+++ library/common/src/main/java/com/google/android/exoplayer2/upstream/HttpDataSource.java
@@ -306,22 +306,51 @@ public interface HttpDataSource extends DataSource {
      */
     public final Map<String, List<String>> headerFields;
 
-    /** @deprecated Use {@link #InvalidResponseCodeException(int, String, Map, DataSpec)}. */
+    /** The response body. */
+    public final byte[] responseBody;
+
+    /**
+     * @deprecated Use {@link #InvalidResponseCodeException(int, String, Map, DataSpec, byte[])}.
+     */
     @Deprecated
     public InvalidResponseCodeException(
         int responseCode, Map<String, List<String>> headerFields, DataSpec dataSpec) {
-      this(responseCode, /* responseMessage= */ null, headerFields, dataSpec);
+      this(
+          responseCode,
+          /* responseMessage= */ null,
+          headerFields,
+          dataSpec,
+          /* responseBody= */ Util.EMPTY_BYTE_ARRAY);
+    }
+
+    /**
+     * @deprecated Use {@link #InvalidResponseCodeException(int, String, Map, DataSpec, byte[])}.
+     */
+    @Deprecated
+    public InvalidResponseCodeException(
+        int responseCode,
+        @Nullable String responseMessage,
+        Map<String, List<String>> headerFields,
+        DataSpec dataSpec) {
+      this(
+          responseCode,
+          responseMessage,
+          headerFields,
+          dataSpec,
+          /* responseBody= */ Util.EMPTY_BYTE_ARRAY);
     }
 
     public InvalidResponseCodeException(
         int responseCode,
         @Nullable String responseMessage,
         Map<String, List<String>> headerFields,
-        DataSpec dataSpec) {
+        DataSpec dataSpec,
+        byte[] responseBody) {
       super("Response code: " + responseCode, dataSpec, TYPE_OPEN);
       this.responseCode = responseCode;
       this.responseMessage = responseMessage;
       this.headerFields = headerFields;
+      this.responseBody = responseBody;
     }
 
   }